Move chart web components into separate directories
It's probably easier to understand our component structure if we stick
to a consistent one, rather then exporting a bunch of components from
one directory, while other directories have a 1:1 relationship to
components.
