--- services/ui/src/js/components/Chart/index.js
@@ -1,2 +1,0 @@
-export { default as Chart2DHistogram } from "./Chart2DHistogram";
-export { default as ChartHistogram } from "./ChartHistogram";
