+++ services/ui/src/js/components/Chart2DHistogram/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Chart2DHistogram";
