+++ services/ui/src/js/components/ChartElement/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ChartElement";
