+++ services/ui/src/js/components/ChartHistogram/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ChartHistogram";
