+++ services/ui/src/js/helpers.js
