+++ services/ui/src/js/main.js
@@ -1,7 +1,8 @@
 import App from "./App.svelte";
 
-import { Chart2DHistogram, ChartHistogram } from "./components/Chart";
+import Chart2DHistogram from "./components/Chart2DHistogram";
 import ChartContainer from "./components/ChartContainer";
+import ChartHistogram from "./components/ChartHistogram";
 import ChartMap from "./components/ChartMap";
 
 customElements.define("chart-2dhistogram", Chart2DHistogram);
